+++ src/AngleSharp.Css.Tests/Functions/CssColorFunction.cs
@@ -19,6 +19,39 @@ namespace AngleSharp.Css.Tests.Functions
             var color = s.GetColor();
             Assert.AreEqual("rgba(0, 17, 0, 1)", color);
         }
+
+        [Test]
+        public void ParseRgbWithSpacesL4_Issue131()
+        {
+            var html = @"<p style='color: rgb(255 122 127 / 80%)'>Text</p>";
+            var dom = ParseDocument(html);
+            var p = dom.QuerySelector("p");
+            var s = p.GetStyle();
+            var color = s.GetColor();
+            Assert.AreEqual("rgba(255, 122, 127, 0.8)", color);
+        }
+
+        [Test]
+        public void ParseRgbWithSpacesInPercentL4_Issue131()
+        {
+            var html = @"<p style='color: rgb(100% 10% 50% / 0.7)'>Text</p>";
+            var dom = ParseDocument(html);
+            var p = dom.QuerySelector("p");
+            var s = p.GetStyle();
+            var color = s.GetColor();
+            Assert.AreEqual("rgba(255, 26, 128, 0.7)", color);
+        }
+
+        [Test]
+        public void ParseRgbWithSpacesAndNoneL4_Issue131()
+        {
+            var html = @"<p style='color: rgb(100% none 50% / 35%)'>Text</p>";
+            var dom = ParseDocument(html);
+            var p = dom.QuerySelector("p");
+            var s = p.GetStyle();
+            var color = s.GetColor();
+            Assert.AreEqual("rgba(255, 0, 128, 0.35)", color);
+        }
     }
 }